♻️ 코드 컨벤션 및 구조 개선
