+++ types/post.type.ts
@@ -45,12 +45,3 @@ export interface PostInputType {
   representative_place_id?: number;
   isviewd: boolean;
 }
-
-export interface SortButtonProps {
-  label: string;
-  value: SortOption;
-  selected: boolean;
-  onClick: (value: SortOption) => void;
-}
-
-export type SortOption = 'latest' | 'popular' | 'rating' | 'likes';
